Remove get_table_or_404 function to mathesar.utils
